moved struct ssl_decoder_ definition into header file because we need it in ssl.enums.c
